Resume_insight/requirements.txt
streamlit
pdfminer.six
python-docx
fuzzywuzzy
python-Levenshtein